+++ CHANGELOG.md
@@ -72,6 +72,9 @@ your dev-utils dependency for your project to build.
 * New `Spinner` component returns a simple img-based spinner as an animated PNG, available in two
   sizes. Used for the platform-specific `Mask` and `LoadingIndicator` components. Replaces previous
   SVG-based implementations to mitigate rendering performance issues over remote connections.
+* New flags `withCommas` and `omitFourDigitComma` for `fmtNumber()` to customize the treatment of
+commas in number displays.
+
 
 ### 💥 Breaking Changes